Addressing comments and resolve conflicts
