+++ gobblin-modules/gobblin-orc/src/main/java/org/apache/gobblin/writer/GobblinOrcWriterBuilder.java
@@ -26,6 +26,9 @@ import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
 
 
+/**
+ * The WriterBuilder extension to create {@link GobblinOrcWriter} on top of {@link FsDataWriterBuilder}
+ */
 public class GobblinOrcWriterBuilder extends FsDataWriterBuilder<Schema, GenericRecord> {
   public GobblinOrcWriterBuilder() {
   }
